+++ src/test/java/com/aws/greengrass/shadowmanager/util/ValidatorTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.shadowmanager.ipc;
+package com.aws.greengrass.shadowmanager.util;
 
 import com.aws.greengrass.shadowmanager.exception.InvalidRequestParametersException;
 import com.aws.greengrass.shadowmanager.model.ShadowRequest;
